tui/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-terminal"
version = "0.1.0"
description = "A terminal emulator backed entirely by an LLM"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.80.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
]

[project.scripts]
llm-terminal = "llm_terminal.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["llm_terminal"]
